+++ CHANGELOG.md
@@ -146,6 +146,7 @@
 123. Added Better Exception Handling for MelonLoader Core Initialization.
 124. Added Error Log File Output to Installer.
 125. Added MelonPreferences Auto-Downgrade System to Installer.
+126. Added MelonUtils.ColorToANSI to convert ConsoleColor to ANSI String.
 
 ---